// src/api-client/models/api-response-page-response-user-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PageResponseUserResponse } from './page-response-user-response';

/**
 * 
 * @export
 * @interface ApiResponsePageResponseUserResponse
 */
export interface ApiResponsePageResponseUserResponse {
    /**
     * 
     * @type {number}
     * @memberof ApiResponsePageResponseUserResponse
     */
    'code'?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiResponsePageResponseUserResponse
     */
    'message'?: string;
    /**
     * 
     * @type {PageResponseUserResponse}
     * @memberof ApiResponsePageResponseUserResponse
     */
    'result'?: PageResponseUserResponse;
}